# internal/openapi/swagger.yaml
basePath: /api
...
info:
  contact: {}
  description: HTTP API for the AWG Manager daemon (tunnels, routing, system).
  title: AWG Manager API
  version: "1.0"
paths:
...
  /system/update/apply:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Apply system update
      tags:
      - update
  /system/update/changelog:
    get:
      parameters:
      - description: From version
        in: query
        name: from
        type: string
      - description: To version
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Update changelog
      tags:
      - update
  /system/update/check:
    get:
      parameters:
      - description: Force refresh from upstream
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Update check
      tags:
      - update
...
  /terminal/install:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Install terminal (ttyd)
      tags:
      - terminal
  /terminal/start:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Start terminal
      tags:
      - terminal
  /terminal/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Terminal status
      tags:
      - terminal
  /terminal/stop:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Stop terminal
      tags:
      - terminal
  /terminal/ws:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: WebSocket upgrade
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Terminal WebSocket
      tags:
      - terminal
  /test/connectivity:
    get:
      parameters:
      - description: Tunnel id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Connectivity check
      tags:
      - test
  /test/ip:
    get:
      parameters:
      - description: Tunnel id
        in: query
        name: id
        type: string
      - description: Check service id
        in: query
        name: service
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: IP leak check
      tags:
      - test
  /test/ip/services:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - CookieAuth: []
      summary: IP check services
      tags:
      - test
  /test/speed:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Speed test (sync)
      tags:
      - test
  /test/speed/servers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Speed test servers
      tags:
      - test
  /test/speed/stream:
    get:
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Speed test stream
      tags:
      - test
...
securityDefinitions:
  CookieAuth:
    description: Session cookie set by POST /auth/login. Omit for public routes.
    in: cookie
    name: awg_session
    type: apiKey
swagger: "2.0"